Extract SupportedMimeType to decorator-free runtime-common subpath
Move SupportedMimeType enum from router.ts to a standalone
supported-mime-type.ts in @cardstack/runtime-common. This subpath has
no transitive imports of decorator-using modules, making it safe for
Playwright's TypeScript loader which doesn't support experimentalDecorators.

- Add packages/runtime-common/supported-mime-type.ts
- Remove packages/software-factory/src/mime-types.ts (local copy)
- All software-factory files import from @cardstack/runtime-common/supported-mime-type
- Barrel re-export and router.ts updated to use the new source
