+++ packages/runtime-common/index.ts
@@ -309,7 +309,7 @@ export const isNode =
   Object.prototype.toString.call((globalThis as any).process) ===
   '[object process]';
 
-export { SupportedMimeType } from './router';
+export { SupportedMimeType } from './supported-mime-type';
 export {
   isUrlLike,
   VirtualNetwork,
