+++ packages/runtime-common/router.ts
@@ -47,26 +47,7 @@ function formatUnknownError(error: unknown): string {
 
 export type Method = 'GET' | 'QUERY' | 'POST' | 'PATCH' | 'DELETE' | 'HEAD';
 
-/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
-export enum SupportedMimeType {
-  CardJson = 'application/vnd.card+json',
-  CardSource = 'application/vnd.card+source',
-  FileMeta = 'application/vnd.card.file-meta+json',
-  DirectoryListing = 'application/vnd.api+json',
-  RealmInfo = 'application/vnd.api+json',
-  Mtimes = 'application/vnd.api+json',
-  Permissions = 'application/vnd.api+json',
-  Session = 'application/json',
-  EventStream = 'text/event-stream',
-  HTML = 'text/html',
-  JSONAPI = 'application/vnd.api+json',
-  JSON = 'application/json',
-  CardDependencies = 'application/json',
-  CardTypeSummary = 'application/json',
-  OctetStream = 'application/octet-stream',
-  All = '*/*',
-}
-/* eslint-enable @typescript-eslint/no-duplicate-enum-values */
+export { SupportedMimeType } from './supported-mime-type';
 
 function isHTTPMethod(method: unknown): method is Method {
   if (typeof method !== 'string') {
