+++ packages/software-factory/scripts/lib/factory-agent.ts
@@ -1,4 +1,4 @@
-import { SupportedMimeType } from '../../src/mime-types';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 import { createBoxelRealmFetch } from '../../src/realm-auth';
