+++ packages/software-factory/src/factory-target-realm.ts
@@ -4,7 +4,7 @@ import {
   iconURLFor,
   getRandomBackgroundURL,
 } from '@cardstack/runtime-common/realm-display-defaults';
-import { SupportedMimeType } from '@cardstack/runtime-common/router';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 import {
   getAccessibleRealmTokens,
