+++ packages/software-factory/src/harness/api.ts
@@ -1,6 +1,6 @@
 import { resolve } from 'node:path';
 
-import { SupportedMimeType } from '../mime-types';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 import {
   readPreparedTemplateMetadata,
