+++ packages/software-factory/tests/factory-entrypoint.integration.test.ts
@@ -4,7 +4,7 @@ import { createServer } from 'node:http';
 import { resolve } from 'node:path';
 import { module, test } from 'qunit';
 
-import { SupportedMimeType } from '../src/mime-types';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 const packageRoot = resolve(__dirname, '..');
 const stickyNoteFixture = readFileSync(
