+++ packages/software-factory/tests/factory-target-realm.test.ts
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 
-import { SupportedMimeType } from '../src/mime-types';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 import { FactoryEntrypointUsageError } from '../src/factory-entrypoint-errors';
 import {
