+++ packages/software-factory/tests/factory-tool-executor.test.ts
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 
-import { SupportedMimeType } from '../src/mime-types';
+import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type';
 
 import type { AgentAction, ToolResult } from '../scripts/lib/factory-agent';
 import {
